requirements.txt
fastapi
uvicorn
pydantic
langchain>=0.1.0
langchain-core
langchain-community
langchain-openai
chromadb
python-dotenv
rank_bm25
presidio-analyzer
presidio-anonymizer
spacy
deepeval
pytest
httpx